+++ packages/ozone/tests/3p-labeler.test.ts
@@ -23,7 +23,7 @@ describe('labels from 3p labelers', () => {
 
   beforeAll(async () => {
     network = await TestNetwork.create({
-      dbPostgresSchema: 'ozone_admin_third_party_labeler',
+      dbPostgresSchema: 'ozone_admin_third_party_labeler_main',
     })
     ozone = network.ozone
 
@@ -34,7 +34,7 @@ describe('labels from 3p labelers', () => {
       plcUrl: network.plc.url,
       signingKey: ozoneKey,
       serverDid: ozoneDid,
-      dbPostgresSchema: `ozone_admin_third_party_labeler`,
+      dbPostgresSchema: `ozone_admin_third_party_labeler_third_party`,
       dbPostgresUrl: ozone.ctx.cfg.db.postgresUrl,
       appviewUrl: network.bsky.url,
       appviewDid: network.bsky.ctx.cfg.serverDid,