+++ packages/dev-env/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @atproto/dev-env
 
+## 0.3.63
+
+### Patch Changes
+
+- Updated dependencies [[`a4b528e5f`](https://github.com/bluesky-social/atproto/commit/a4b528e5f51c8bfca56b293b0059b88d138ec421), [`2e7aa211d`](https://github.com/bluesky-social/atproto/commit/2e7aa211d2cbc629899c7f87f1713b13b932750b), [`90399c859`](https://github.com/bluesky-social/atproto/commit/90399c85955301babc689c293bd3e7e1a94505a3)]:
+  - @atproto/api@0.13.17
+  - @atproto/pds@0.4.72
+  - @atproto/bsky@0.0.95
+  - @atproto/ozone@0.1.56
+
+## 0.3.62
+
+### Patch Changes
+
+- Updated dependencies [[`24423fc2d`](https://github.com/bluesky-social/atproto/commit/24423fc2dd394c99a29dbe4419b356090ef19546), [`bac9be2d3`](https://github.com/bluesky-social/atproto/commit/bac9be2d3ec904d1f984a871f43cf89aca17289d), [`48d08a469`](https://github.com/bluesky-social/atproto/commit/48d08a469f75837e3b7e879d286d12780440b8b8), [`bac9be2d3`](https://github.com/bluesky-social/atproto/commit/bac9be2d3ec904d1f984a871f43cf89aca17289d), [`bac9be2d3`](https://github.com/bluesky-social/atproto/commit/bac9be2d3ec904d1f984a871f43cf89aca17289d), [`561431fe4`](https://github.com/bluesky-social/atproto/commit/561431fe4897e81767dc768e9a31020d09bf86ff)]:
+  - @atproto/pds@0.4.71
+  - @atproto/syntax@0.3.1
+  - @atproto/ozone@0.1.55
+  - @atproto/api@0.13.16
+  - @atproto/lexicon@0.4.3
+  - @atproto/bsky@0.0.94
+  - @atproto/bsync@0.0.9
+  - @atproto/sync@0.1.6
+  - @atproto/xrpc-server@0.7.3
+
 ## 0.3.61
 
 ### Patch Changes